backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
minio==7.2.11
redis==5.2.1
torch==2.5.1
transformers==4.48.0
hdbscan==0.8.41
scikit-learn==1.6.0
faiss-cpu==1.9.0.post1
numpy==1.26.4
requests==2.32.3